+++ source/blender/editors/uvedit/uvedit_parametrizer.c
@@ -2672,8 +2672,8 @@ static PBool p_abf_matrix_invert(PAbfSystem *sys, PChart *chart)
 		}
 
 		for (i = 0; i < ninterior; i++) {
-			sys->lambdaPlanar[i] += nlGetVariable(0, i);
-			sys->lambdaLength[i] += nlGetVariable(0, ninterior + i);
+			sys->lambdaPlanar[i] += (float)nlGetVariable(0, i);
+			sys->lambdaLength[i] += (float)nlGetVariable(0, ninterior + i);
 		}
 	}
 
@@ -4561,7 +4561,7 @@ void param_pack(ParamHandle *handle, float margin, bool do_rotate)
 		box->index = i; /* warning this index skips PCHART_NOPACK boxes */
 		
 		if (margin > 0.0f)
-			area += sqrtf(box->w * box->h);
+			area += (double)sqrtf(box->w * box->h);
 	}
 	
 	if (margin > 0.0f) {
